+++ changelog.md
@@ -14,6 +14,7 @@ http://semver.org/spec/v2.0.0.html).
 - ntdll.NtQueryTimerResolution
 - sample cnt_apis.py
 - sample totp_dlg.pyw
+- userenv.GetUserProfileDirectory
 
 ### Changed